backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
gunicorn==23.0.0
packaging==25.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
yapf==0.43.0
